requirements.txt
vncdotool
opencv-python
pytesseract
watchdog
colorlog
imutils
configargparse
image
scikit-image
mysql-connector-python
twisted
gpxdata
flask
requests
s2sphere
imagehash